requirements.txt
absl-py==0.8.0
astor==0.8.0
atari-py==0.2.6
cached-property==1.5.2
certifi==2020.11.8
chardet==3.0.4
cloudpickle==1.6.0
cycler==0.10.0
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.33.2
gym==0.17.3
h5py==3.1.0
idna==2.10
importlib-metadata==3.0.0
joblib==0.17.0
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.4
Pillow==8.0.1
protobuf==3.14.0
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
scipy==1.5.4
six==1.15.0
stable-baselines==2.10.0
tensorboard==1.15.0
tensorflow-gpu==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
